How to control 4 servo motors using arduino
I do not own this code it is the property of Fabio Cardoso this code is to be used for non-commercial purposes here is a link of his video where this code was used https://www.youtube.com/watch?v=px3Bc8dniYg. I modified this code so it can control 4 servo motors I used this code for non-commercial purposes.