How to control 4 servo motors using arduino
I did not create this code nor do I own it. This code is the property of Fabio Cardoso the reason this code is here in the first place is for non-commercial purposes only. Here is a Fabio's video where this code was used https://www.youtube.com/watch?v=px3Bc8dniYg.